Fix for Versioneer
Versioneer requires that the package be built in a valid git
repository. `pip wheel .` copies the package to a temporary
directory then builds the wheel. The copying breaks the
repository state of the package because of the broken relative
path in the .git file of a submodule. It is not enough to
convert to an absolute path since linux builds use docker
which isolates the repository under a different rootpath.

To remedy this, the package(submodule) is replaced by a git
repository with a .git directory.

Note that it was not enough to copy the original git_dir to
replace the .git file since the contents of the .git/config
point to a worktree via a relative path.

Fixes #11