NF: make submodule .git directory for versioneer
pip will copy the directory it installs from before running the install
/ wheel build.  If the directory is a submodule, then the copy breaks
the connection with the containing project, so git commands to get the
version will fail.

Fix is to re-create the .git directory for submodules.

Version of fix in https://github.com/matthew-brett/multibuild/pull/12
with thanks to Hassan Kibirige.