pass MB_PYTHON_VERSION to docker run in build_multilinux
It is useful sometimes (e.g. in config.sh's pre_build') to do stuff based on
the current python version. The problem is that when running inside the docker
container the variable MB_PYTHON_VERSION is passed on only as PYTHON_VERSION
without the MB_ prefix, but in the macOS environment the latter is not
present. Thus, one needs to check both PYTHON_VERSION and MB_PYTHON_VERSION.
Better to pass both to the docker container, so one can jut check the value
of MB_PYTHON_VERSION on both mac and linux.

Note that the same hack is used in the 'install_run' function, where both
PYTHON_VERSION and MB_PYTHON_VERSION are passed through to docker run.